Diagramas e classe java modificados

+++ src/main/java/br/com/agrodrone/domain/AreaAgricola.java
@@ -4,14 +4,14 @@ package br.com.agrodrone.domain;
  * Define uma área agrícola monitorada pela cooperativa.
  */
 public class AreaAgricola {
-    private Long id;
+    private int id;
     private String localizacao;
-    private Double tamanho;       // em hectares
+    private double tamanho;       // em hectares
     private String tipoCultivo;   // ex.: soja, milho
 
     public AreaAgricola() {}
 
-    public AreaAgricola(Long id, String localizacao, Double tamanho, String tipoCultivo) {
+    public AreaAgricola(int id, String localizacao, double tamanho, String tipoCultivo) {
         this.id = id;
         this.localizacao = localizacao;
         this.tamanho = tamanho;
@@ -19,12 +19,12 @@ public class AreaAgricola {
     }
 
     // Getters e setters
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
     public String getLocalizacao() { return localizacao; }
     public void setLocalizacao(String localizacao) { this.localizacao = localizacao; }
-    public Double getTamanho() { return tamanho; }
-    public void setTamanho(Double tamanho) { this.tamanho = tamanho; }
+    public double getTamanho() { return tamanho; }
+    public void setTamanho(double tamanho) { this.tamanho = tamanho; }
     public String getTipoCultivo() { return tipoCultivo; }
     public void setTipoCultivo(String tipoCultivo) { this.tipoCultivo = tipoCultivo; }
 }
